Separate class in files
